pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "brazil-daily-news"
version = "0.1.0"
description = "巴西经贸信息周报自动生成系统"
requires-python = ">=3.11"
dependencies = [
    "PyYAML",
    "requests",
    "beautifulsoup4",
    "openai>=1.0.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
]

[project.scripts]
brazil-news = "brazil_daily_news.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
